+++ test/utilities_tests.jl
@@ -15,12 +15,6 @@ using QuickHeaps
     @test QuickHeaps.is_one_based_unit_range(2:4) == false
     @test QuickHeaps.is_one_based_unit_range(1:2:5) == false
 
-    let A = rand(Float32, 2)
-        @test_deprecated QuickHeaps.to_eltype(A, 11)
-        @test QuickHeaps.to_eltype(A, 11) isa Float32
-        @test QuickHeaps.to_eltype(A, π) isa Float32
-    end
-
     @test QuickHeaps.in_range(0, 3) == false
     @test QuickHeaps.in_range(1, 3) == true
     @test QuickHeaps.in_range(3, 3) == true